requirements.txt
flask
pymongo
bcrypt
PyJWT
python-dotenv
requests
unidecode
elasticsearch
flask_swagger_ui